==> Pipeline/4-Summarization.ipynb <==
## Import Libraries

import pandas as pd
import nltk
from nltk.tokenize import sent_tokenize, word_tokenize
from nltk.corpus import stopwords
from nltk.probability import FreqDist
from string import punctuation
from collections import defaultdict
from heapq import nlargest
import numpy as np

## Load Data

jc_number = pd.read_csv(r"C:\Users\Olga\Becode_Olga\KPMG-Team-3\data_csv\C\clustered\jc_119_0_cluster.csv")

## Summarize the Text for the Target Cluster with the Predefined Keywords

def get_context(text, freq_words):
    sents = sent_tokenize(text)
    freq = FreqDist(freq_words)
    ranking = defaultdict(int)

    for i, sent in enumerate(sents):
        for w in word_tokenize(sent.lower()):
            if w in freq:
                ranking[i] += freq[w]
                
    sents_idx = nlargest(6, ranking, key=ranking.get)
    return [sents[j].replace('\n', ' ')  for j in sorted(sents_idx)]

dic_words = {
            'cluster_0' : ['premie', 'jaarlijkse', 'december', 'betaald', 'betaalbaar', 'collectieve', 'arbeidsovereenkomst', 'betreffende', 'commission', 'paritair', 'bedienden']
            }
search_words = dic_words ['cluster_0']         

jc_number['row_number'] = np.arange(len(jc_number))

## Get text in a column "summary_nl"
class_0 = jc_number.loc[jc_number['class'] == 0]
class_0['summary_nl'] = jc_number.apply( lambda row : get_context(row['nl_text'], search_words), axis = 1)

## Get text as summary example 
rows =[]
for r in class_0['row_number']:
    rows.append(r)

for i in rows:
    text = jc_number['nl_text'].iloc[i]  
    context = get_context(text, search_words)
    print('\n'.join(context))
    print('\n*********************************************\n')

## Save Dataframe with Summarization to Excel

class_0.to_excel(r"C:\Users\Olga\Becode_Olga\KPMG-Team-3\data_csv\C\summary\jc_119_0_summary.xlsx", engine='xlsxwriter')
